improved technician profile backend structure
