+++ apps/server/src/modules/technicians/index.ts
@@ -4,9 +4,9 @@ export {
   type ITechniciansRepository,
   type CreateTechnicianData,
   type UpdateTechnicianData,
+  type TechnicianProfile,
+  type TechnicianProfileRow,
 } from './technicians.repository.js';
-export { TechniciansService, type ITechniciansService, type TechnicianProfile } from './technicians.service.js';
+export { TechniciansService, type ITechniciansService } from './technicians.service.js';
 export { TechniciansController } from './technicians.controller.js';
-export { default as techniciansRoutes } from './technicians.routes.js';
-export { default as technicianProfileRoutes } from './technician-profile.routes.js';
-
+export { techniciansRoutes, technicianProfileRoutes } from './technicians.routes.js';
